extract transformation.go
